requirements.txt
mlx>=0.4.0
numpy>=1.24.0
Pillow>=10.0.0
PyYAML>=6.0
tqdm>=4.65.0
anthropic>=0.39.0
